Create udef_arp_qgis.py
This is the UDef_ARP.py but modified for QGIS. Some differences to note: The QGIS plugin imports Qt widgets from qgis.PyQt. New statements have been added to work with the QgsMapLayerComboBox widget. For example the selectRaster function and other arguments to set the styles and parameters for these widgets. Within each processing function, a try except logic is added to read the input raster layers as either QGIS layers or text. Also an if statement is added that enables users to render results within a QGIS project. Finally, a separate class object is added at the bottom to represent the plugin itself. Within the class are functions that enable QGIS to recognize and initialize the python script as a plugin.